track connect + live reload for logging
